remove ephem requirement
No other files in this repository mention ephem, so presumably this rather
heavyweight library is a stale dependency that is no longer needed.  Also, since
pip seems to want to build a shared library from source when installing ephem on
Linux and Darwin, including ephem as a dependency to django_celery_beat has the
unintended consequence of requiring Python development headers.