Fix unstable unittest
The origin unittest use loop to confirm task m1 will not
be blocked by task m2. This is not a stable way, the test
sometimes passed, sometimes failed. Because m1 sechdued at
every 3s, it may or may not schdeuled at after m2' start_time
1s, but current time is before m2' start_time 2s, which make
the test failed, but not as expected.